Refactored much of the training pipeline
- almost restructured the training script so that the best model for
each scenario is pushed. Previously, the best tuned and untuned models
for each scenario were pushed instead. This resulted in a need for me to
delete two of the models in the model registry, so that the best two
models would remain.

- simplified the makefile and the training script so that all the models
get trained in one shot.

- added functionality that ensures that at the beginning of each
training run, all prior experiments and all the models from the model
registry get cleared.

- switched much of the code in the training pipeline to a more
functional style because I feel that my use of classes was no longer
justifiable.