Implemented a change in the preprocessing script
The change ensures that when the preprocessing script is run, any
existing training data is deleted. This change is long overdue, as the
previous implementation simply skipped the generation of training data
if a file of the same name is present in the designated directory. This
made it so that I had to delete outdated versions manually.

With that being said, the real solution to this problem would be to
check whether the file is outdated and delete it only when it is clearly
out of date.