Identified a strategy for ensuring that training data is only deleted when new data is available
I will use the cleaning process. When the data is to be cleaned, a
function will check any existing cleaned data and delete it if it's old
and out of date relative to the current state of Lyft's S3 bucket.

I also intend to refactor the entire training script, because it is very
hard to follow.