I seem to have at least partially resolved the issue with model naming
The way the code around model naming was written was needlessly complex.
However, there's now an issue that causes the training process to be
stalled while logging details about the xgboost model to Comet. It
doesn't seem to have anything to do with the size of the data set, as I
never had this issue before, and I the issue is not solved by creating
smaller training datasets